+++ Tarea7/TareaReact/src/components/BoardsList.tsx
