+++ Tarea7/TareaReact/src/main.tsx
@@ -16,7 +16,8 @@ import App from './App.tsx'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom'
-import { BoardsList } from "./BoardsList"
+import { BoardsList } from "./components/BoardsList.tsx"
+import { Toaster } from 'react-hot-toast'
 import './index.css'  
 
 // Inicializamos el cliente de React Query
@@ -31,6 +32,7 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
           <Route path="/board/:boardId" element={<App />} />
           <Route path="*" element={<Navigate to="/" />} />
         </Routes>
+        <Toaster position="top-right" />
         <ReactQueryDevtools initialIsOpen={false} />
       </BrowserRouter>
     </QueryClientProvider>
